backend/Piper/requirements.txt
numpy
soundfile
piper-tts